@@ -6,7 +6,7 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(LOXCPP_ROOT ${PROJECT_SOURCE_DIR})
 set(LOXCPP_SRCS_DIR "${LOXCPP_ROOT}/src/")
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++17")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
 
 set(LOXCPP_SRCS)
 list(APPEND LOXCPP_SRCS
@@ -20,12 +20,12 @@ list(APPEND LOXCPP_SRCS
     #${LOXCPP_SRCS_DIR}/LoxClass.cpp
     #${LOXCPP_SRCS_DIR}/LoxFunction.cpp
     #${LOXCPP_SRCS_DIR}/LoxInstance.cpp
-    #${LOXCPP_SRCS_DIR}/Parser.cpp
+    ${LOXCPP_SRCS_DIR}/Parser.cpp
     #${LOXCPP_SRCS_DIR}/Resolver.cpp
-    #${LOXCPP_SRCS_DIR}/Return.cpp
+    ${LOXCPP_SRCS_DIR}/Return.cpp
     ${LOXCPP_SRCS_DIR}/RuntimeError.cpp
     ${LOXCPP_SRCS_DIR}/Scanner.cpp
-    #${LOXCPP_SRCS_DIR}/Stmt.cpp
+    ${LOXCPP_SRCS_DIR}/Stmt.cpp
     ${LOXCPP_SRCS_DIR}/Token.cpp
     ${LOXCPP_SRCS_DIR}/main.cpp
 )
