@@ -1,14 +1,14 @@
 #ifndef RETURN_H
 #define RETURN_H
 
+#include <stdexcept>
+#include <string>
+#include <variant>
+
+
 namespace Lox {
 
-class Return : public std::runtime_error {
- public:
-  const std::string value;
-
-  Return::Return(std::string value);
-};
+class Return;
 
 }  // namespace Lox
 
