@@ -18,143 +18,132 @@ class Stmt {};
 template <class T>
 class Visitor;
 
-/*template <typename T>
-class Visitor<T> {
-
-public:
-    virtual T visitBlockStmt(Block stmt);
-    virtual T visitClassStmt(Class stmt);
-    virtual T visitExpressionStmt(Expression stmt);
-    virtual T visitFunctionStmt(Function stmt);
-    virtual T visitIfStmt(If stmt);
-    virtual T visitPrintStmt(Print stmt);
-    virtual T visitReturnStmt(Return stmt);
-    virtual T visitVarStmt(Var stmt);
-    virtual T visitWhileStmt(While stmt);
-
-};*/
-
 
 template <typename T>
 class Block : public Stmt<T> {
  public:
-  Block(std::vector<Stmt<T>> statements);
+  Block(const std::vector<Stmt<T>>& statements);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
   const std::vector<Stmt<T>> statements;
 };
 
-
+/*
 template <typename T>
 class Class : public Stmt<T> {
  public:
   Class(
-      Token,
-      lox::expr::Expr::Variable superclass,
-      std::vector<Stmt::Function> methods);
+      const Token& name,
+      const lox::expr::Variable<T>& superclass,
+      const std::vector<Function<T>>& methods);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
   const Token name;
-  const lox::expr::Expr::Variable superclass;
-  const std::vector<Stmt::Function> methods;
+  const lox::expr::Variable<T> superclass;
+  const std::vector<Function<T>> methods;  // TODO: check &
+};
+*/
+
+template <class T>
+class Expression : public Stmt<T> {
+ public:
+  Expression(const lox::expr::Expr<T>& expression);
+
+  T accept(const Visitor<T>& visitor);
+
+  const lox::expr::Expr<T> expression;
 };
 
 
 template <class T>
-class Expression : public Stmt {
+class Function : public Stmt<T> {
  public:
-  Expression(lox::expr::Expr expression);
+  Function(
+      const Token& name,
+      const std::vector<Token>& params,
+      const std::vector<Stmt<T>>& body);
 
-  T accept(Visitor<T> visitor);
-
-  const lox::expr::Expr expression;
-};
-
-
-template <class T>
-class Function : public Stmt {
- public:
-  Function(Token name, std::vector<Token> params, std::vector<Stmt> body);
-
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
   const Token name;
   const std::vector<Token> params;
-  const std::vector<Stmt> body;
+  const std::vector<Stmt<T>> body;
 };
 
 
 template <class T>
-class If : public Stmt {
+class If : public Stmt<T> {
  public:
-  If(lox::expr::Expr condition, Stmt thenBranch, Stmt elseBranch);
+  If(const lox::expr::Expr<T>& condition,
+     const Stmt<T>& thenBranch,
+     const Stmt<T>& elseBranch);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
-  const lox::expr::Expr condition;
-  const Stmt thenBranch;
-  const Stmt elseBranch;
+  const lox::expr::Expr<T> condition;
+  const Stmt<T> thenBranch;
+  const Stmt<T> elseBranch;
 };
 
 
 template <class T>
-class Print : public Stmt {
+class Print : public Stmt<T> {
  public:
-  Print(lox::expr::Expr expression);
+  Print(const lox::expr::Expr<T>& expression);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
-  const lox::expr::Expr expression;
+  const lox::expr::Expr<T> expression;
 };
 
 
 template <class T>
-class Return : public Stmt {
+class Return : public Stmt<T> {
  public:
-  Return(Token keyword, lox::expr::Expr value);
+  Return(const Token& keyword, const lox::expr::Expr<T>& value);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
   const Token keyword;
-  const lox::expr::Expr value;
+  const lox::expr::Expr<T> value;
 };
 
 
 template <class T>
-class Var : public Stmt {
+class Var : public Stmt<T> {
  public:
-  Var(Token name, lox::expr::Expr initializer);
+  Var(const Token& name, const lox::expr::Expr<T>& initializer);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
   const Token name;
-  const lox::expr::Expr initializer;
+  const lox::expr::Expr<T> initializer;
 };
 
 
 template <class T>
-class While : public Stmt {
+class While : public Stmt<T> {
  public:
-  While(lox::expr::Expr condition, Stmt body);
+  While(const lox::expr::Expr<T>& condition, const Stmt<T>& body);
 
-  T accept(Visitor<T> visitor);
+  T accept(const Visitor<T>& visitor);
 
-  const lox::expr::Expr condition;
-  const Stmt body;
+  const lox::expr::Expr<T> condition;
+  const Stmt<T> body;
 };
 
 
 template <class T>
-T accept(Visitor<T> visitor);
+T accept(const Visitor<T>& visitor);
 
 
 template <class T>
 class Visitor : public Stmt<T> {
  public:
   virtual T visitBlockStmt(const Block<T>& stmt) = 0;
-  virtual T visitClassStmt(const Class<T>& stmt) = 0;
+  // virtual T visitClassStmt(const Class<T>& stmt) = 0;
   virtual T visitExpressionStmt(const Expression<T>& stmt) = 0;
   virtual T visitFunctionStmt(const Function<T>& stmt) = 0;
   virtual T visitIfStmt(const If<T>& stmt) = 0;
